Address Alanna's nit picks

+++ src/interest/views.py
@@ -5,35 +5,35 @@ from api.utils import failure_response
 from api.utils import success_response
 from interest.models import Interest
 from rest_framework import generics
-from rest_framework import status as s
+from rest_framework import status
 
 from .controllers.create_interest_controller import CreateInterestController
 from .serializers import InterestSerializer
 
 
-class InterestAllView(generics.GenericAPIView):
+class InterestsView(generics.GenericAPIView):
     serializer_class = InterestSerializer
     permission_classes = api_settings.CONSUMER_PERMISSIONS
 
     def get(self, request):
-        """ Get all interests """
+        """Get all interests."""
         interests = Interest.objects.all()
         return success_response(self.serializer_class(interests, many=True).data)
 
     def post(self, request):
-        """ Create an interest """
+        """Create an interest."""
         return CreateInterestController(
             request=request, serializer=self.serializer_class
         ).process()
 
     def delete(self, request):
-        """ Delete an interest """
-        try:
-            body = json.loads(request.body)
-            name = body.get("name")
-            assert name
-        except:
-            return failure_response("POST body is misformatted", s.HTTP_400_BAD_REQUEST)
+        """Delete an interest."""
+        body = json.loads(request.body)
+        name = body.get("name")
+        if name is None:
+            return failure_response(
+                "POST body is misformatted", status.HTTP_400_BAD_REQUEST
+            )
         interest = Interest.objects.filter(name=name)
         if interest:
             interest[0].delete()
